build it right

+++ scheduler.c
@@ -6,7 +6,7 @@
 struct job* job_enq(struct command *cmd) {
 	struct job *j;
 	j = malloc(sizeof(struct job));
-	j->job_status = 0;
+	j->job_status = JS_INIT;
 	j->job_init_priority = 3;
 	j->job_cur_priority = 0;
 	j->job_waiting = 0;
@@ -16,11 +16,13 @@ struct job* job_enq(struct command *cmd) {
 	j->job_args[j->job_args_len] = '/0';
 	
 	LOCK(&sc->sc_lock);
-	LIST_INSERT_HEAD(&sc->sc_jobs, j, job_entries);
 	j->job_id = sc->sc_next_job_id++;
+	
+	job_loader(j);
+	LIST_INSERT_HEAD(&sc->sc_jobs, j, job_entries);
 	UNLOCK(&sc->sc_lock);
 
-	printf("job_enq: jobid %d, jobargs %s\n", j->job_id, j->job_args);
+	DEBUG("job_enq: jobid %d, jobargs %s\n", j->job_id, j->job_args);
 	return j;
 }
 
@@ -28,10 +30,10 @@ int job_deq(struct command *cmd) {
 	struct job *j, *found;
 	int32_t job_id;
 	job_id = atoi(cmd->args);
-	printf("job_deq: cmd->args %s\n",cmd->args);
+	DEBUG("job_deq: cmd->args %s\n",cmd->args);
 	found = NULL;
 	int ret = 1;
-	printf("job_deq: remove job %d\n", job_id);
+	DEBUG("job_deq: remove job %d\n", job_id);
 	LOCK(&sc->sc_lock);
 	LIST_FOREACH(j, &sc->sc_jobs, job_entries) {
 		if (j->job_id == job_id) {
@@ -63,7 +65,7 @@ int job_loader(struct job *job) {
 				argc++;
 		}
 		argc++;
-		printf("job_loader: argc %d\n", argc);
+		DEBUG("job_loader: argc %d\n", argc);
 		
 		argv = (char **)malloc(sizeof(char *) * argc);
 		
@@ -72,42 +74,46 @@ int job_loader(struct job *job) {
 		i = 0;
 		token = strtok (job->job_args, delimiters);
 		argv[i] = token;
-		printf("job_loader: token %s\n", token);
+		DEBUG("job_loader: token %s\n", token);
 		while (token != NULL) {
-			token = strtok (NULL, delimiters);
+			token = strtok(NULL, delimiters);
 			i++;
 			argv[i] = token;
-			printf("job_loader: token %s\n", token);
+			DEBUG("job_loader: token %s\n", token);
 		}
 		
-		printf("job_loader: child execve %s\n");
+		DEBUG("job_loader: child execve %s\n");
 		ret = execve(argv[0], argv, NULL);
 		if (ret < 0 ) {
-			printf("job_loader: job %d load fail\n", job->job_id);
+			DEBUG("job_loader: job %d load fail\n", job->job_id);
 		}
-		printf("job_loader: child exit\n");
+		DEBUG("job_loader: child exit\n");
+		free(argv);
 		// need or not?
-		// exit(0);
+		exit(0);
 	} else {
-		printf("job_loader: parent kill\n");
+		DEBUG("job_loader: parent kill\n");
 		job->job_pid = pid;
 		kill(pid, SIGSTOP);
+		job->job_status = JS_READY;
 	}
 }
 
 void job_switch(struct job *current, struct job *next) {
-	printf("job_switch: \n");
+	DEBUG("job_switch: \n");
 	
 	if (current != NULL && next != NULL) {
-		printf("job_switch: from %d to %d\n", current->job_pid, next->job_pid);
+		PRINTF("job_switch: from %d to %d\n", current->job_pid, next->job_pid);
 	}
 	
-	if (current != NULL) {
+	if (current != NULL && current->job_pid != 0) {
 		kill(current->job_pid, SIGSTOP);
+		current->job_status = JS_WAIT;
 	}
-	if (next != NULL){
+	if (next != NULL && next->job_pid != 0){
 		current = next;
-		kill(current->job_pid, SIGCONT);		
+		kill(current->job_pid, SIGCONT);
+		current->job_status = JS_RUNNING;		
 	}
 }
 
@@ -116,15 +122,36 @@ void wait_current_job() {
 	int ret;
 	
 	LOCK(&sc->sc_lock);
-	if (sc->sc_current == NULL) {
-		return;
+	if (sc->sc_current == NULL || sc->sc_current->job_pid == 0) {
+		goto unlock;
 	}
 	ret = waitpid(sc->sc_current->job_pid, &status, WNOHANG);
-	printf("clean_finishjob: waitpid %d get %d\n", sc->sc_current->job_pid, ret);
+	DEBUG("clean_finishjob: waitpid %d get %d\n", sc->sc_current->job_pid, ret);
 	if (ret == sc->sc_current->job_pid) {
-		printf("clean_finishjob: current job %d finished\n", ret);
-		LIST_REMOVE(sc->sc_current, job_entries);
+		DEBUG("clean_finishjob: current job %d finished\n", ret);
+		PRINTF("wait_current_job: job finish  job %d,  pid %d", sc->sc_current->job_id, sc->sc_current->job_pid);
+		sc->sc_current->job_status = JS_DONE;
 		sc->sc_current = NULL;
+	}
+unlock:
+	UNLOCK(&sc->sc_lock);
+}
+
+void clean_done_job() {
+	struct job *j, *done_job;
+	LOCK(&sc->sc_lock);
+	LIST_FOREACH(j, &sc->sc_jobs, job_entries) {
+		if (j->job_status == JS_DONE) {
+			PRINTF("clean_done_job: clean job %d,  pid %d", j->job_id, j->job_pid);
+			LIST_REMOVE(j, job_entries);
+			LIST_INSERT_HEAD(&sc->sc_done_jobs, j, job_entries);
+		}
+	}
+	
+	while (!LIST_EMPTY(&sc->sc_done_jobs)) {
+		j = LIST_FIRST(&sc->sc_done_jobs);
+		LIST_REMOVE(j, job_entries);
+		free(j);
 	}
 	UNLOCK(&sc->sc_lock);
 }
@@ -144,9 +171,12 @@ void do_schedule() {
 	sc->sc_scheduler_last_runtime = cur_time;
 	
 	LIST_FOREACH(j, &sc->sc_jobs, job_entries) {
+		if (!(j->job_status == JS_READY || j->job_status == JS_WAIT))
+			continue;
 		if (sc->sc_current == NULL) {
 			sc->sc_current = j;
 		}
+
 		j->job_waiting += diff_ms;
 		if (j->job_waiting >= 100) {
 			if (0 != j->job_cur_priority) {
@@ -165,22 +195,26 @@ void do_schedule() {
 		sc->sc_current->job_waiting = 0;
 		job_switch(sc->sc_current, next);
 		sc->sc_current = next;
+		if (sc->sc_current->job_pid == 0) {
+			PRINTF("!!!!!!!!!!!!!!!!!!!!!\n");
+			PRINTF("do_schedule: current job_id is null %d", sc->sc_current);
+		}
 	}
 	UNLOCK(&sc->sc_lock);
 }
 
 static int socket_listen() {
-	printf("socket_listen: call socket_listen\n");
+	PRINTF("socket_listen: call socket_listen\n");
 	int ret = -1;
 	struct sockaddr_in *listener_addr;
 	if (sc->sc_listener_sock != -1) {
-		printf("socket_listen: already listening\n");
+		PRINTF("socket_listen: already listening\n");
 		return ret;
 	}
 	sc->sc_listener_sock = socket(AF_INET, SOCK_STREAM, 0);
 	
 	if (sc->sc_listener_sock == -1) {
-		printf("socket_listen: create socket fail\n");
+		PRINTF("socket_listen: create socket fail\n");
 	}
 	
 	listener_addr = &sc->sc_listener_addr;
@@ -190,16 +224,16 @@ static int socket_listen() {
 	
 	ret = bind(sc->sc_listener_sock, (struct sockaddr *)listener_addr, sizeof(struct sockaddr));
 	if (ret == -1) {
-		printf("socket_listen: bind fail\n");
+		PRINTF("socket_listen: bind fail\n");
 		return ret;
 	}
 	
 	ret = listen(sc->sc_listener_sock, 10);
 	if (ret == -1) {
-		printf("socket_listen: listen fail\n");
+		PRINTF("socket_listen: listen fail\n");
 		return ret;
 	}
-	printf("socket_listen: listening....\n");
+	PRINTF("socket_listen: listening....\n");
 	return ret;
 }
 
@@ -209,24 +243,24 @@ struct command * recvcommand() {
 	struct sockaddr_in client_addr;
 	struct command *cmd;
 	if (sc->sc_listener_sock == -1) {
-		printf("recvcommand: sc_listener_sock not ready\n");
+		DEBUG("recvcommand: sc_listener_sock not ready\n");
 		return NULL;
 	}
 	while (1) {
 		sin_size = sizeof(struct sockaddr_in);
 		if((client_fd = accept(sc->sc_listener_sock, (struct sockaddr *)&client_addr, &sin_size)) == -1) {
-			//printf("recvcommand: accept error");
+			//PRINTF("recvcommand: accept error");
 			continue;
 		}
-		printf("listener: got connection\n");
+		DEBUG("listener: got connection\n");
 		memset(buf, 0, sizeof(buf));
 		numbytes = recv(client_fd, buf, MAXDATASIZE, 0);
 		if(numbytes == -1) {
-			printf("recvcommand: recv fail\n");
+			PRINTF("recvcommand: recv fail\n");
 		}
 		if (numbytes) {
 			buf[numbytes] = '/0';
-			printf("receive: %s\n", buf + CMDHEADSIZE);
+			DEBUG("receive: %s\n", buf + CMDHEADSIZE);
 		}
 		cmd = (struct command *)buf;
 		do_req(client_fd, cmd);
@@ -234,6 +268,7 @@ struct command * recvcommand() {
 }
 
 static void sig_timer(int i) {
+	clean_done_job();
 	wait_current_job();
 	do_schedule();
 }
@@ -243,6 +278,7 @@ void init_sc() {
 	
 	sc->sc_listener_sock = -1;
 	LIST_INIT(&sc->sc_jobs);
+	LIST_INIT(&sc->sc_done_jobs);
 	LIST_INIT(&jobs_head);
 	LOCK_INIT(&sc->sc_lock);
 	
@@ -255,7 +291,7 @@ static void send_reply(int32_t sock, char *buf, int buf_size) {
 	int ret;
 	ret = send(sock, buf, buf_size, 0);
 	if (ret == -1) {
-		printf("send_reply: send fail\n");
+		DEBUG("send_reply: send fail\n");
 		return ret;
 	}
 }
@@ -264,12 +300,13 @@ int do_enq_req(int32_t sock, struct command * req) {
 	char buf[MAXDATASIZE];
 	struct job *j;
 	j = job_enq(req);
-	job_loader(j);
+	//job_loader(j);
 	memset(buf, 0, sizeof(buf));
 	sprintf(buf, "%d", j->job_id);
 	send_reply(sock, buf, strlen(buf));
 	shutdown(sock, 2);
 	close(sock);
+	//DEBUG("close enq_req socket");
 }
 
 int do_deq_req(int32_t sock, struct command * req) {
@@ -287,13 +324,15 @@ int do_deq_req(int32_t sock, struct command * req) {
 int do_stat_req(int32_t sock, struct command * req) {
 	struct job *j;
 	char buf[MAXDATASIZE];
-	printf("do_stat_req: \n");
+	DEBUG("do_stat_req: \n");
 	
+	LOCK(&sc->sc_lock);
 	LIST_FOREACH(j, &sc->sc_jobs, job_entries) {
 		memset(buf, 0, sizeof(buf));
-		snprintf(buf, sizeof(buf), "%d %s \n",j->job_id, j->job_args);
+		snprintf(buf, sizeof(buf), "%d %s %d \n",j->job_id, j->job_args, j->job_status);
 		send_reply(sock, buf, strlen(buf));
 	}
+	UNLOCK(&sc->sc_lock);
 	shutdown(sock, 2);
 	close(sock);
 }
@@ -310,7 +349,7 @@ int do_req(int32_t sock, struct command * req) {
 		do_stat_req(sock, req);
 		break;
 	default:
-		printf("not support command\n");
+		DEBUG("not support command\n");
 	}
 }
